debian: change rules to build for python3 and gtk3
Add the needed stuff to build for python3 and gtk3.

Please notice that we still need python 2.x, as some
test scripts depend on it.

NOTE: for now, keeping kust debian-sid with gcc, as it
complains that gtk-3.0 package was not found on Xenial.

As, for now, this is only for testing, we don't need to
build for everything anyway. So, let's focus on just one
setting.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>